+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-- Removed the api key prefix validation for non enterprise Dfuse editions.
+- Removed the api key prefix validation.
 
 ## 0.3.17 (October 20, 2020)