+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Removed the api key prefix validation for non enterprise dfuse editions.
+
 - Improved error notification for GraphQL stream when the server rejects a message that is too big. Now, the stream will received an error and will close instead of staying alive while the socket is down.
 
 ## 0.3.17 (October 20, 2020)