+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Added a default authentication endpoint for the dfuse Community Edition.
+
 - Removed the api key prefix validation.
 
 - Improved error notification for GraphQL stream when the server rejects a message that is too big. Now, the stream will received an error and will close instead of staying alive while the socket is down.